move functionality into outbound transport manager
