+++ aries_cloudagent/conductor.py
@@ -11,10 +11,7 @@ wallet.
 import asyncio
 import hashlib
 import logging
-import time
-from typing import Union
 
-from .delivery_queue import DeliveryQueue
 from .admin.base_server import BaseAdminServer
 from .admin.server import AdminServer
 from .config.default_context import ContextBuilder
@@ -27,42 +24,14 @@ from .protocols.connections.manager import ConnectionManager, ConnectionManagerE
 from .messaging.responder import BaseResponder
 from .messaging.task_queue import TaskQueue
 from .stats import Collector
-from .transport.pack_format import PackWireFormat
 from .transport.inbound.manager import InboundTransportManager
 from .transport.inbound.message import InboundMessage
 from .transport.inbound.session import InboundSession
+from .transport.outbound.base import OutboundDeliveryError
 from .transport.outbound.manager import OutboundTransportManager
 from .transport.outbound.message import OutboundMessage
 
 LOGGER = logging.getLogger(__name__)
-
-
-class QueuedOutboundMessage:
-    STATE_NEW = "new"
-    STATE_PENDING = "pending"
-    STATE_ENCODE = "encode"
-    STATE_DELIVER = "deliver"
-    STATE_RETRY = "retry"
-    STATE_DONE = "done"
-
-    def __init__(
-        self,
-        context: InjectionContext,
-        message: OutboundMessage,
-        target,
-        transport_id: str,
-    ):
-        self.context = context
-        self.endpoint = target and target.endpoint
-        self.error: Exception = None
-        self.message = message
-        self.payload: Union[str, bytes] = None
-        self.retries = 0
-        self.retry_at: float = None
-        self.state = self.STATE_NEW
-        self.target = target
-        self.task: asyncio.Task = None
-        self.transport_id: str = transport_id
 
 
 class Conductor:
@@ -86,14 +55,9 @@ class Conductor:
         self.admin_server = None
         self.context: InjectionContext = None
         self.context_builder = context_builder
-        self.delivery_queue: TaskQueue = None
         self.dispatcher: Dispatcher = None
         self.inbound_transport_manager: InboundTransportManager = None
-        self.outbound_buffer = []
-        self.outbound_event: asyncio.Event = None
-        self.outbound_new = []
         self.outbound_transport_manager: OutboundTransportManager = None
-        self.undelivered_queue: DeliveryQueue = None
 
     async def setup(self):
         """Initialize the global request context."""
@@ -102,28 +66,17 @@ class Conductor:
 
         self.dispatcher = Dispatcher(context)
 
-        # Fetch stats collector, if any
-        collector = await context.inject(Collector, required=False)
-
         # Register all inbound transports
         self.inbound_transport_manager = InboundTransportManager(
             context, self.inbound_message_router
         )
         await self.inbound_transport_manager.setup()
 
-        # FIXME add config options
-        self.delivery_queue = TaskQueue(max_active=50)
-        self.outbound_event = asyncio.Event()
-
         # Register all outbound transports
-        self.outbound_transport_manager = OutboundTransportManager(collector)
-        outbound_transports = context.settings.get("transport.outbound_configs") or []
-        for outbound_transport in outbound_transports:
-            try:
-                self.outbound_transport_manager.register(outbound_transport)
-            except Exception:
-                LOGGER.exception("Unable to register outbound transport")
-                raise
+        self.outbound_transport_manager = OutboundTransportManager(
+            context, self.dispatcher.run_task
+        )
+        await self.outbound_transport_manager.setup()
 
         # Admin API
         if context.settings.get("admin.enabled"):
@@ -146,8 +99,8 @@ class Conductor:
                 LOGGER.exception("Unable to register admin server")
                 raise
 
-        self.context = context
-
+        # Fetch stats collector, if any
+        collector = await context.inject(Collector, required=False)
         if collector:
             # add stats to our own methods
             collector.wrap(
@@ -155,7 +108,7 @@ class Conductor:
                 (
                     # "inbound_message_router",
                     "outbound_message_router",
-                    "create_inbound_session",
+                    # "create_inbound_session",
                 ),
             )
             collector.wrap(self.dispatcher, "handle_message")
@@ -168,6 +121,8 @@ class Conductor:
                     "find_message_connection",
                 ),
             )
+
+        self.context = context
 
     async def start(self) -> None:
         """Start the agent."""
@@ -192,9 +147,7 @@ class Conductor:
             LOGGER.exception("Unable to start outbound transports")
             raise
 
-        loop = asyncio.get_event_loop()
-        loop.create_task(self.process_outbound())
-        # loop.create_task(self.log_status())
+        # asyncio.get_event_loop().create_task(self.log_status())
 
         # Start up Admin server
         if self.admin_server:
@@ -314,64 +267,6 @@ class Conductor:
                 f"{e:>4} pack  {p:>4} send  {t:>4} out"
             )
 
-    async def queue_processing(self, session: InboundSession):
-        """
-        Interact with undelivered queue to find applicable messages.
-
-        Args:
-            session: The inbound session
-        """
-        if (
-            session
-            and session.reply_mode
-            and not session.closed
-            and self.undelivered_queue
-        ):
-            for key in session.reply_verkeys:
-                if not isinstance(key, str):
-                    key = key.value
-                if self.undelivered_queue.has_message_for_key(key):
-                    for (
-                        undelivered_message
-                    ) in self.undelivered_queue.inspect_all_messages_for_key(key):
-                        # pending message. Transmit, then kill single_response
-                        if session.select_outbound(undelivered_message):
-                            LOGGER.debug(
-                                "Sending Queued Message via inbound connection"
-                            )
-                            self.undelivered_queue.remove_message_for_key(
-                                key, undelivered_message
-                            )
-                            await session.send(undelivered_message)
-
-    def select_inbound_session(self, outbound: OutboundMessage):
-        # if outbound.reply_session_id:
-
-        # if inbound and inbound.receipt.
-
-        #  if the message has multiple targets, we cannot direct return unless
-        #  one of the targets has keys that match the inbound message (!)
-        # if an
-
-        accepted = False
-
-        # try open inbound sessions first, preferring the same session ID
-        # FIXME if outbound target is set, need to compare to inbound keys
-
-        if not outbound.target:
-            inbound_session = self.inbound_sessions.get(outbound.reply_session_id)
-            if inbound_session:
-                accepted, retry = inbound_session.accept_response(outbound)
-
-            if not accepted:
-                for inbound_session in self.inbound_sessions.values():
-                    if inbound_session.session_id != outbound.reply_session_id:
-                        accepted, retry = inbound_session.accept_response(outbound)
-                        break
-
-        if accepted:
-            LOGGER.debug("Returned message to socket %s", inbound_session.session_id)
-
     async def outbound_message_router(
         self,
         context: InjectionContext,
@@ -390,6 +285,7 @@ class Conductor:
         # if inbound and inbound.direct_response:
         #     if outbound.reply_to_verkey
 
+        # FIXME - use dispatch task
         # always populate connection targets using provided context
         if not outbound.target and not outbound.target_list and outbound.connection_id:
             mgr = ConnectionManager(context)
@@ -401,126 +297,12 @@ class Conductor:
                 LOGGER.exception("Error preparing outbound message for transmission")
                 return
 
-        targets = [outbound.target] if outbound.target else (outbound.target_list or [])
-        transport_id = None
-        out_mgr = self.outbound_transport_manager
-        for target in targets:
-            endpoint = target.endpoint
-            transport_id = out_mgr.get_running_transport_for_endpoint(endpoint)
-            if transport_id:
-                break
-        if not transport_id:
-
+        try:
+            self.outbound_transport_manager.deliver(context, outbound)
+        except OutboundDeliveryError:
             # Add message to outbound queue, indexed by key
             # if self.undelivered_queue:
             #     self.undelivered_queue.add_message(message)
 
             LOGGER.warning("Cannot queue message for delivery, no supported transport")
             return  # drop message
-
-        queued = QueuedOutboundMessage(context, outbound, target, transport_id)
-        self.outbound_new.append(queued)
-        self.outbound_event.set()
-
-    async def process_outbound(self):
-        """Continually watch the outbound buffer and send to transports."""
-
-        while True:
-            # if self.stopping .. break
-
-            self.outbound_event.clear()
-            loop_time = time.perf_counter()
-            upd_buffer = []
-
-            for queued in self.outbound_buffer:
-                if queued.state == QueuedOutboundMessage.STATE_DONE:
-                    if queued.error:
-                        LOGGER.exception(
-                            "Outbound message could not be delivered",
-                            exc_info=queued.error,
-                        )
-                    continue  # remove from buffer
-
-                deliver = False
-
-                if queued.state == QueuedOutboundMessage.STATE_PENDING:
-                    deliver = True
-                elif queued.state == QueuedOutboundMessage.STATE_RETRY:
-                    if queued.retry_at < loop_time:
-                        queued.retry_at = None
-                        deliver = True
-
-                if deliver:
-                    queued.state = QueuedOutboundMessage.STATE_DELIVER
-                    self.deliver_queued_message(queued)
-
-                upd_buffer.append(queued)
-
-            new_messages = self.outbound_new.copy()
-            self.outbound_new = []
-            for queued in new_messages:
-
-                if queued.message.enc_payload:
-                    queued.payload = queued.message.enc_payload
-                    queued.state = QueuedOutboundMessage.STATE_DELIVER
-                    self.deliver_queued_message(queued)
-                else:
-                    queued.state = QueuedOutboundMessage.STATE_ENCODE
-                    self.encode_queued_message(queued)
-
-                upd_buffer.append(queued)
-
-            self.outbound_buffer = upd_buffer
-            await self.outbound_event.wait()
-
-    def encode_queued_message(self, queued: QueuedOutboundMessage) -> asyncio.Task:
-        transport = self.outbound_transport_manager.get_transport(queued.transport_id)
-        wire_format = transport.wire_format or PackWireFormat()
-        queued.task = self.dispatcher.run_task(
-            wire_format.encode_message(
-                queued.context,
-                queued.message.payload,
-                queued.target.recipient_keys,
-                queued.target.routing_keys,
-                queued.target.sender_key,
-            ),
-            lambda task, exc_info: self.finished_encode(queued, task, exc_info),
-        )
-        return queued.task
-
-    def finished_encode(
-        self, queued: QueuedOutboundMessage, task: asyncio.Task, exc_info=None
-    ):
-        if exc_info:
-            queued.error = exc_info
-            queued.state = QueuedOutboundMessage.STATE_DONE
-        else:
-            queued.payload = task.result()
-            queued.state = QueuedOutboundMessage.STATE_PENDING
-        queued.task = None
-        self.outbound_event.set()
-
-    def deliver_queued_message(self, queued: QueuedOutboundMessage) -> asyncio.Task:
-        transport = self.outbound_transport_manager.get_transport(queued.transport_id)
-        queued.task = self.delivery_queue.run(
-            transport.handle_message(queued.payload, queued.endpoint),
-            lambda task, exc_info: self.finished_deliver(queued, task, exc_info),
-        )
-        return queued.task
-
-    def finished_deliver(
-        self, queued: QueuedOutboundMessage, task: asyncio.Task, exc_info=None
-    ):
-        """Clean up a closed session."""
-        if exc_info:
-            queued.error = exc_info
-            if queued.retries < 5:
-                queued.state = QueuedOutboundMessage.STATE_RETRY
-                queued.retry_at = time.perf_counter() + 10
-            else:
-                queued.state = QueuedOutboundMessage.STATE_DONE
-        else:
-            queued.error = None
-            queued.state = QueuedOutboundMessage.STATE_DONE
-        queued.task = None
-        self.outbound_event.set()
